+++ app/lib/fragmentmodeler/laneLabelHandling/index.js
@@ -1,0 +1,6 @@
+import LaneLabelHandler from "./laneLabelHandler";
+
+export default {
+  __init__: ["laneLabelHandler"],
+  laneLabelHandler: ["type", LaneLabelHandler],
+};
