--- app/lib/fragmentmodeler/taskLabelHandling/index.js
@@ -1,6 +1,0 @@
-import TaskLabelHandler from "./taskLabelHandler";
-
-export default {
-  __init__: ["taskLabelHandler"],
-  taskLabelHandler: ["type", TaskLabelHandler],
-};
